[FEATURE] Add the possibility to disable the autopopup of the consent modal.
This is useful on privacy policy pages if the cookie consent is mandatory and cannot be
closed before a consent decision was made. If you have it full-screen on mobile devices
then you probably want to enable the user to read the privacy policy before they take an
informed decision.